Spring

+++ tomatofarm/src/main/java/teksystems/tomatofarm/database/entity/User.java
@@ -2,10 +2,14 @@ package teksystems.tomatofarm.database.entity;
 
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.annotation.CreatedDate;
+import teksystems.tomatofarm.database.dao.UserRoleDAO;
 
 import javax.persistence.*;
 import java.util.Date;
+import java.util.List;
+import java.util.Set;
 
 @Getter
 @Setter
@@ -20,7 +24,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "first_name")
     private String firstName;
@@ -37,4 +41,10 @@ public class User {
     @Column(name = "create_date")
     @Temporal(TemporalType.TIMESTAMP)
     private Date createDate = new Date();
+
+    //TODO: create attribute - list of strings of user's roles
+    //use @OneToMany annotation?
+//    private Set<UserRole> userRoles; // <-- maybe this
+//    private List<UserRole> userRoles;// <-- or maybe this since all we need is strings
+//    Java is saying basic attribute type can't be container though
 }
